requirements.txt
aiohttp
aiolimiter
python-dotenv
tqdm